introduce new module Parser
