+++ cdm_reader_mapper/mdf_reader/utils/filereader.py
@@ -2,22 +2,19 @@
 
 from __future__ import annotations
 
-import ast
-import csv
 import logging
 import os
 
 import pandas as pd
 
-from itertools import zip_longest
 
 from .. import properties
-from ..schemas import schemas
 from .utilities import validate_path, process_textfilereader
-from .utilities import convert_dtypes, remove_boolean_values
+from .utilities import remove_boolean_values
 
-from .convert_and_decode import Converters, Decoders, convert_and_decode
+from .convert_and_decode import convert_and_decode
 from .validators import validate
+from .parser import parse_fixed_width, parse_delimited, Parser
 
 
 def _apply_multiindex(df: pd.DataFrame, length) -> pd.DataFrame:
@@ -30,87 +27,16 @@ def _apply_multiindex(df: pd.DataFrame, length) -> pd.DataFrame:
     return df
 
 
-def _validate_sentinel(i: int, line: str, sentinel: str) -> bool:
-    return line.startswith(sentinel, i)
-
-
-def _get_index(section, order, length):
-    if length == 1:
-        return section
-    return (order, section)
-
-
-def _get_ignore(section_dict) -> bool:
-    ignore = section_dict.get("ignore", False)
-    if isinstance(ignore, str):
-        ignore = ast.literal_eval(ignore)
-    return bool(ignore)
-
-
-def _parse_fixed_width(
-    line: str,
-    i: int,
-    header: dict,
-    compiled_elements: list,
-    sections: list,
-    out: dict,
-) -> int:
-    section_length = header.get("length", properties.MAX_FULL_REPORT_WIDTH)
-    delimiter = header.get("delimiter")
-    sentinel = header.get("sentinel")
-
-    bad_sentinel = sentinel is not None and not _validate_sentinel(i, line, sentinel)
-    k = i + section_length
-
-    for index, na_value, field_length, ignore in compiled_elements:
-        if isinstance(index, tuple):
-            in_sections = index[0] in sections
-        else:
-            in_sections = index in sections
-
-        missing = True
-
-        j = i if bad_sentinel else i + field_length
-        if j > k:
-            missing = False
-            j = k
-
-        if not ignore and in_sections:
-            value = line[i:j]
-            if not value.strip() or value == na_value:
-                value = True
-            if i == j and missing:
-                value = False
-            out[index] = value
-
-        if delimiter and line[j : j + len(delimiter)] == delimiter:
-            j += len(delimiter)
-
-        i = j
-
-    return i
-
-
-def _parse_delimited(
-    line: str,
-    i: int,
-    order: str,
-    header: dict,
-    elements: dict,
-    olength: int,
-    out: dict,
-) -> int:
-    delimiter = header["delimiter"]
-    fields = next(csv.reader([line[i:]], delimiter=delimiter))
-
-    for name, value in zip_longest(elements.keys(), fields):
-        out[_get_index(name, order, olength)] = (
-            value.strip() if value is not None else None
-        )
-        if value is not None:
-            i += len(value)
-
-    return i
+def _apply_or_chunk(data, func, func_args=[], func_kwargs={}, **kwargs):
+    if not isinstance(data, pd.io.parsers.TextFileReader):
+        return func(data, *func_args, **func_kwargs)
+    return process_textfilereader(
+        data,
+        func,
+        func_args,
+        func_kwargs,
+        **kwargs,
+    )
 
 
 class FileReader:
@@ -143,118 +69,16 @@ class FileReader:
         self.year_end = year_end
         self.ext_table_path = ext_table_path
 
-        logging.info("READING DATA MODEL SCHEMA FILE...")
-        if ext_schema_path or ext_schema_file:
-            self.schema = schemas.read_schema(
-                ext_schema_path=ext_schema_path, ext_schema_file=ext_schema_file
-            )
-        else:
-            self.schema = schemas.read_schema(imodel=imodel)
-
-        parsing_order = self.schema["header"].get("parsing_order")
-        sections_ = [x.get(y) for x in parsing_order for y in x]
-        self.orders = [y for x in sections_ for y in x]
-        self.olength = len(self.orders)
-
-        self._build_compiled_specs_and_convertdecode()
-
         self.pd_kwargs = {}
         self.xr_kwargs = {}
 
         self.sections = None
         self.encoding = None
 
-    def _build_compiled_specs_and_convertdecode(self):
-        compiled_specs = []
-        disable_reads = []
-        dtypes = {}
-        converter_dict = {}
-        converter_kwargs = {}
-        decoder_dict = {}
-
-        for order in self.orders:
-            section = self.schema["sections"][order]
-            header = section["header"]
-            elements = section["elements"]
-
-            disable_read = header.get("disable_read", False)
-            if disable_reads:
-                disable_reads.append(order)
-
-            compiled_elements = []
-            for name, meta in elements.items():
-                index = _get_index(name, order, self.olength)
-                ignore = _get_ignore(meta)
-
-                compiled_elements.append(
-                    (
-                        index,
-                        meta.get("missing_value"),
-                        meta.get("field_length", properties.MAX_FULL_REPORT_WIDTH),
-                        ignore,
-                    )
-                )
-
-                if disable_read:
-                    continue
-
-                if ignore:
-                    continue
-
-                ctype = meta.get("column_type")
-                dtype = properties.pandas_dtypes.get(ctype)
-
-                if dtype:
-                    dtypes[index] = dtype
-
-                conv_func = Converters(ctype).converter()
-                if conv_func:
-                    converter_dict[index] = conv_func
-
-                conv_kwargs = {
-                    k: meta.get(k)
-                    for k in properties.data_type_conversion_args.get(ctype, [])
-                }
-                if conv_kwargs:
-                    converter_kwargs[index] = conv_kwargs
-
-                encoding = meta.get("encoding")
-                if encoding:
-                    dec_func = Decoders(ctype, encoding).decoder()
-                    if dec_func:
-                        decoder_dict[index] = dec_func
-
-            compiled_specs.append(
-                (
-                    order,
-                    header,
-                    elements,
-                    compiled_elements,
-                    header.get("format") == "delimited",
-                )
-            )
-
-        self.dtypes, self.parse_dates = convert_dtypes(dtypes)
-
-        self.disable_reads = disable_reads
-
-        self.convert_decode = {
-            "converter_dict": converter_dict,
-            "converter_kwargs": converter_kwargs,
-            "decoder_dict": decoder_dict,
-        }
-
-        self.compiled_specs = compiled_specs
-
-    def _apply_or_chunk(self, data, func, func_args=[], func_kwargs={}, **kwargs):
-        if not isinstance(data, pd.io.parsers.TextFileReader):
-            return func(data, *func_args, **func_kwargs)
-        return process_textfilereader(
-            data,
-            func,
-            func_args,
-            func_kwargs,
-            **kwargs,
+        self.parser = Parser(
+            imodel=imodel,
+            ext_schema_path=ext_schema_path,
+            ext_schema_file=ext_schema_file,
         )
 
     def _read_line(self, line: str) -> dict:
@@ -267,17 +91,17 @@ class FileReader:
             elements,
             compiled_elements,
             is_delimited,
-        ) in self.compiled_specs:
+        ) in self.parser.compiled_specs:
             if header.get("disable_read"):
                 out[order] = line[i : properties.MAX_FULL_REPORT_WIDTH]
                 continue
 
             if is_delimited:
-                i = _parse_delimited(
-                    line, i, order, header, elements, self.olength, out
+                i = parse_delimited(
+                    line, i, order, header, elements, self.parser.olength, out
                 )
             else:
-                i = _parse_fixed_width(
+                i = parse_fixed_width(
                     line, i, header, compiled_elements, self.sections, out
                 )
 
@@ -309,7 +133,7 @@ class FileReader:
         col = df.columns[0]
         records = df[col].map(self._read_line)
         df = pd.DataFrame.from_records(records)
-        return _apply_multiindex(df, self.olength)
+        return _apply_multiindex(df, self.parser.olength)
 
     def _apply_schema(
         self,
@@ -318,19 +142,19 @@ class FileReader:
         data = self._open_pandas(data)
         data = convert_and_decode(
             data,
-            converter_dict=self.convert_decode["converter_dict"],
-            converter_kwargs=self.convert_decode["converter_kwargs"],
-            decoder_dict=self.convert_decode["decoder_dict"],
+            converter_dict=self.parser.convert_decode["converter_dict"],
+            converter_kwargs=self.parser.convert_decode["converter_kwargs"],
+            decoder_dict=self.parser.convert_decode["decoder_dict"],
         )
         data = self._select_years(data)
         mask = validate(
             data,
             imodel=self.imodel,
             ext_table_path=self.ext_table_path,
-            schema=self.schema,
-            disables=self.disable_reads,
+            schema=self.parser.schema,
+            disables=self.parser.disable_reads,
         )
-        data = remove_boolean_values(data, self.dtypes)
+        data = remove_boolean_values(data, self.parser.dtypes)
         return data, mask
 
     def _open_with_pandas(
@@ -345,7 +169,7 @@ class FileReader:
             skip_blank_lines=False,
             **kwargs,
         )
-        return self._apply_or_chunk(
+        return _apply_or_chunk(
             to_parse,
             self._apply_schema,
             makecopy=False,
@@ -363,11 +187,8 @@ class FileReader:
         if open_with == "netcdf":
             raise NotImplementedError
         elif open_with == "pandas":
-            if encoding is None:
-                encoding = self.schema["header"].get("encoding", "utf-8")
-
-            self.sections = sections
-            self.encoding = encoding
+            self.sections = sections or self.parser.orders
+            self.encoding = encoding or self.parser.encoding
             self.pd_kwargs = {
                 "encoding": encoding,
                 "chunksize": chunksize,
